Modify server run script
Noticed on Raspberry pi killing if there was no existing
session wasn't working as there was no running server.
So now checking first if there's a session

+++ server/run.sh
@@ -1,4 +1,9 @@
 # cron with
 # @reboot path/to/run.sh
-tmux kill-session -t journal-server
-tmux new-session -d -s 'journal-server' 'python main.py'
+tmux has-session -t 'journal-server' 2>/dev/null
+if [ $? != 0 ]; then
+  tmux new-session -d -s 'journal-server' 'python main.py'
+else
+  tmux kill-session -t journal-server
+  tmux new-session -d -s 'journal-server' 'python main.py'
+fi
